Refactor: move metrics related files into dir metrics/
