+++ openraft/src/core/admin.rs
@@ -19,13 +19,13 @@ use crate::error::LearnerIsLagging;
 use crate::error::LearnerNotFound;
 use crate::error::MissingNodeInfo;
 use crate::error::NotAllowed;
+use crate::metrics::RemoveTarget;
 use crate::raft::AddLearnerResponse;
 use crate::raft::ChangeMembers;
 use crate::raft::ClientWriteResponse;
 use crate::raft::EntryPayload;
 use crate::raft::RaftRespTx;
 use crate::raft_types::LogIdOptionExt;
-use crate::replication_metrics::RemoveTarget;
 use crate::versioned::Updatable;
 use crate::LogId;
 use crate::Membership;
