+++ openraft/src/core/mod.rs
@@ -43,6 +43,7 @@ use crate::error::ForwardToLeader;
 use crate::error::InitializeError;
 use crate::error::NotAllowed;
 use crate::metrics::RaftMetrics;
+use crate::metrics::ReplicationMetrics;
 use crate::raft::AddLearnerResponse;
 use crate::raft::Entry;
 use crate::raft::EntryPayload;
@@ -52,7 +53,6 @@ use crate::raft::VoteResponse;
 use crate::raft_types::LogIdOptionExt;
 use crate::replication::ReplicaEvent;
 use crate::replication::ReplicationStream;
-use crate::replication_metrics::ReplicationMetrics;
 use crate::storage::RaftSnapshotBuilder;
 use crate::versioned::Versioned;
 use crate::vote::Vote;
