+++ openraft/src/core/replication.rs
@@ -9,12 +9,12 @@ use crate::core::ReplicationState;
 use crate::core::SnapshotState;
 use crate::core::State;
 use crate::error::AddLearnerError;
+use crate::metrics::UpdateMatchedLogId;
 use crate::raft::AddLearnerResponse;
 use crate::raft::RaftRespTx;
 use crate::replication::RaftEvent;
 use crate::replication::ReplicaEvent;
 use crate::replication::ReplicationStream;
-use crate::replication_metrics::UpdateMatchedLogId;
 use crate::storage::Snapshot;
 use crate::summary::MessageSummary;
 use crate::versioned::Updatable;
