+++ openraft/src/metrics/wait.rs
@@ -1,110 +1,18 @@
-//! Raft metrics for observability.
-//!
-//! Applications may use this data in whatever way is needed. The obvious use cases are to expose
-//! these metrics to a metrics collection system like Prometheus. Applications may also
-//! use this data to trigger events within higher levels of the parent application.
-//!
-//! Metrics are observed on a running Raft node via the `Raft::metrics()` method, which will
-//! return a stream of metrics.
-
+use core::time::Duration;
 use std::collections::BTreeSet;
-use std::sync::Arc;
 
-use serde::Deserialize;
-use serde::Serialize;
-use thiserror::Error;
 use tokio::sync::watch;
-use tokio::time::Duration;
 use tokio::time::Instant;
 
-use crate::core::EffectiveMembership;
 use crate::core::State;
-use crate::error::Fatal;
-use crate::raft_types::LogIdOptionExt;
-use crate::replication_metrics::ReplicationMetrics;
-use crate::versioned::Versioned;
+use crate::metrics::RaftMetrics;
+use crate::raft::RaftTypeConfig;
 use crate::LogId;
+use crate::LogIdOptionExt;
 use crate::MessageSummary;
-use crate::RaftTypeConfig;
-
-/// A set of metrics describing the current state of a Raft node.
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RaftMetrics<C: RaftTypeConfig> {
-    pub running_state: Result<(), Fatal<C::NodeId>>,
-
-    /// The ID of the Raft node.
-    pub id: C::NodeId,
-
-    // ---
-    // --- data ---
-    // ---
-    /// The current term of the Raft node.
-    pub current_term: u64,
-
-    /// The last log index has been appended to this Raft node's log.
-    pub last_log_index: Option<u64>,
-
-    /// The last log index has been applied to this Raft node's state machine.
-    pub last_applied: Option<LogId<C::NodeId>>,
-
-    /// The id of the last log included in snapshot.
-    /// If there is no snapshot, it is (0,0).
-    pub snapshot: Option<LogId<C::NodeId>>,
-
-    // ---
-    // --- cluster ---
-    // ---
-    /// The state of the Raft node.
-    pub state: State,
-
-    /// The current cluster leader.
-    pub current_leader: Option<C::NodeId>,
-
-    /// The current membership config of the cluster.
-    pub membership_config: Arc<EffectiveMembership<C>>,
-
-    // ---
-    // --- replication ---
-    // ---
-    /// The metrics about the leader. It is Some() only when this node is leader.
-    pub leader_metrics: Option<Versioned<ReplicationMetrics<C::NodeId>>>,
-}
-
-impl<C: RaftTypeConfig> MessageSummary for RaftMetrics<C> {
-    fn summary(&self) -> String {
-        format!("Metrics{{id:{},{:?}, term:{}, last_log:{:?}, last_applied:{:?}, leader:{:?}, membership:{}, snapshot:{:?}, replication:{}",
-            self.id,
-            self.state,
-            self.current_term,
-            self.last_log_index,
-            self.last_applied,
-            self.current_leader,
-            self.membership_config.summary(),
-            self.snapshot,
-            self.leader_metrics.as_ref().map(|x| x.summary()).unwrap_or_default(),
-        )
-    }
-}
-
-impl<C: RaftTypeConfig> RaftMetrics<C> {
-    pub(crate) fn new_initial(id: C::NodeId) -> Self {
-        Self {
-            running_state: Ok(()),
-            id,
-            state: State::Follower,
-            current_term: 0,
-            last_log_index: None,
-            last_applied: None,
-            current_leader: None,
-            membership_config: Arc::new(EffectiveMembership::default()),
-            snapshot: None,
-            leader_metrics: None,
-        }
-    }
-}
 
 // Error variants related to metrics.
-#[derive(Debug, Error)]
+#[derive(Debug, thiserror::Error)]
 pub enum WaitError {
     #[error("timeout after {0:?} when {1}")]
     Timeout(Duration, String),
